Clarify the treatment of empty structs or unions (#68)
For the base calling convention, empty structs and unions are ignored by
C compilers (if supported as a non-standard language extension). This is
not the case for C++.

Clarify the comment about empty structs (or unions) being ignored in the
floating point calling convention - this refers to how they are
interpreted when 'flattening' structs.