Clarify Usage of Paired Registers
This document does not seem to specify how to split the bits in a 2*XLEN-bit scalar into the pair of XLEN-bit registers. It is not clear which register should contain the low word and which should contain the high word.

This change documents that the low-order XLEN bits go into the lower-numbered register, and the high-order XLEN bits go into the higher-numbered register, which seems to match what GCC and LLVM do.