Adds Linux-specific ABI Section
The specification was previously over-perscriptive in the definitions of
the size, alignment, and signedness of `wchar_t` and `wint_t`.

After some discussion it has been agreed to split these definitions out
from the main specification and into a separate section that only
applies on Linux. This commit introduces that section with the existing
definitions for `wchar_t` and `wint_t`.