Adds Named ABI definitions
This commit adds explicitly named ABI definitions, to match and extend
the conventions used so far by GCC and Clang.

The aim here is to be explicit about what each named ABI means in terms
of which parts of the calling conventions are used, when they are
compatible with different ISAs, and how this maps to C types and sizes.

This Commit defines
- ILP32 (32-bit, no hardware floating-point)
- ILP32F (32-bit, single-precision hardware floating-point)
- ILP32D (32-bit, double-precision hardware floating-point)
- ILP32E (32-bit, no hardware floating point, RVE registers only)
- LP64 (64-bit, no hardware floating-point)
- LP64F (64-bit, single-precision hardware floating-point)
- LP64D (64-bit, double-precision hardware floating-point)
- LP64Q (64-bit, quad-precision hardware floating-point)