Add DWARF register number information for CSRs
There are situations where it might be necessary to reference some
CSRs from the DWARF information.  For example, when unwinding from an
exception handler it could be necessary to reference 'mepc' as the
return address from the exception handler.

Rather than just add DWARF registers numbers for specific CSRs one at
a time as we need them, I'd like to propose that we allocate numbers
for all CSRs in one go.

When thinking how to number the CSRs I see a couple of possible
solutions, first we could allocate 4096 (the maximum number of CSRs)
DWARF register numbers and assign these to CSRs based on the CSR's
number within the specification (Volume II: Privileged Architecture).
The benefit is that the number for a CSR shouldn't need to change in
the future (I think), and adding new CSRs is easy.  The draw back is
that this means allocating a large number of DWARF register numbers,
many of which are currently not used.

A second possibility was to allocate DWARF register numbers
sequentially for the CSRs in the order they currently appear in the
spec.  This would require far fewer DWARF register numbers to be
allocated (maybe just ~250) but if we add more CSRs to the spec later
then these would need to be added to the end of the DWARF register
number list, and things would get very confused.  To try and avoid
this confusion we would probably have to list all of the CSRs in this
document.

In this patch I am suggesting that we take the first approach.  The
CSRs are allocated DWARF numbers in the range 4096 to 8191, this
leaves a big gap between 65 and 4095 that can be used for future
extensions which I have split into two sets.  The first set is for
future standard extensions, and the second set is for custom
extensions.