Disallow combining ILP32E ABI with D ISA extension
As discussed in #125, this denies ILP32E with the D extension. In fact,
it goes slightly further to deny any extensions where the required load
and store alignment is larger than the stack alignment, because that is
the issue at play in #125. This wording should future-proof the psABI
against issues with other extensions as they become ratified.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>