Build PDFs using GitHub actions
On every push, pull request and release creation we build the PDF. On
pushes to master we create a draft release and upload the PDF. On
release creation we upload the PDF to the (non-draft) release.

Any artifacts not associated with a tag (i.e. any draft releases) get
removed after a month, but the 10 most recent artifacts are always kept.