Add support for variant calling conventions for lazily-bound functions
Currently, we save/restore a0-a7 and fa0-fa7 in the glibc and FreeBSD
run-time linker lazy resolvers to avoid clobbering any function
arguments. In order to define the vector calling convention, we need to
specify the behavior of the resolver in the dynamic linker. Vector
registers may have a large size and saving/restoring vector argument
registers may occupy a large portion of stack space during run-time
resolving. It is unreasonable to save/restore all these vector argument
registers in the resolver.

In this patch, we define STO_RISCV_VARIANT_CC, a special symbol
attribute to mark symbols as using a calling convention that passes
arguments in registers other than a0-a7 and fa0-fa7 (such as the vector
calling convention), which run-time linkers can use to eagerly bind such
functions. We also define DT_RISCV_VARIANT_CC, an additional dynamic tag
to indicate there are symbols with the special attribute in the dynamic
symbol table of the object used for PLT-based calls as a quick way for
run-time linkers to skip iterating over the table at load time when no
symbols need to be eagerly bound.

Co-authored-by: Jessica Clarke <jrtc27@jrtc27.com>