Remove [GT]PREL_[IS] internal binutils relocations
These have never been emitted by GCC or Clang, nor has there ever been
assembly syntax in either toolchain to create them. These are fake
internal relocations inside binutils used as part of its implementation
of linker relaxation, and so should not be part of the psABI. They
should be free to be reallocated for psABI uses (which can include
reintroducing them, if there is a need to expose them to assembly and
code generation).

Closes: #160