Fix another syntax highlighting issue
The use of triple backtick doesn't actually end up creating a code block
here, only inline, for both GitHub's and Pandoc's rendering of this, but
it confuses some syntax highlighting which looks for another line with
the closing triple backtick that never comes. Thus, change it to single
backticks to reflect the fact that this is rendered as inline code.