Remove stub compact code model section
This section is currently a draft, so should not feature in any official
publication (and does not currently meet the requirements in the pull
request policy so would never have been added were those in place at the
time). Thus, until a concrete specification for it exists and is deemed
to be a worthwhile addition, we should not talk about it and risk
causing confusion. Especially since, as currently described, it's a new
ABI, not just a new code model, due to the requirement that gp always be
able to be used to access the GOT, which isn't necessarily the case
currently for executables, and is never the case currently for shared
libraries as far as linkers are concerned.