Fix typos for CALL, CALL_PLT and GOT_HI20
`CALL` and `CALL_PLT` relocations refer `auipc` and `jalr` instruction pairs.
`jalr` is not a J-type instruction; it's I-type.

As to `GOT_HI20`, `G` means the offset inside the .got in the other ELF specs.
So the correct expression is not `G+A-P` but `G+GOT+A-P` where `GOT` means
the address of the .got.

I found these errata when I was trying to implement a RISC-V support to my linker.
https://github.com/rui314/mold/commit/e76f7c0d474e667cc0eccac5281709664539b714