Don't differ from the RVI definitions of terms like "frozen"
It's already confusing enough that these definitions are loosely
applied, but having a comment in the spec source that's not even visible
in the spec but re-defines these very important terms just adds an extra
layer of complexity.  Let's just use the RVI terms, that's what everyone
else doesn't do.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>