Require that lazily bound function follow the standard calling convention
glibc's implementation of lazy symbol resolution relies on the function
being bound following some details of the calling convention that are
not included in the register convention, like the direction of stack
growth.  One could imaging building systems that can lazily resolve
functions while only relying on the register convention, but we've got a
decade or so of binaries that don't behave that way and we can't change
those.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>