Adjust the V calling convention to aid lazy binding
The whole point of this VARIANT_CC bit is that we want to be able to use
V registers in _dl_resolve without saving them.  That requires that the
interface between the caller and the callee does not rely on the values
in the V registers.  That's a bit different than how we've histrorically
defined these temporary register clobbers, which is about the the
caller-side visibility into register changes.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>

---

I'm not entirely sure what the right way to describe this one is.  What
I've got here feels a bit clunky and doesn't call out the lazy binding
issue, but at least it's something.  It also only applies to the V
register calling convention, as it's not clear if the goal here is to
retroactively make this apply to X and F.