Open the relocation number more than 255
ELF64 format has support up to 32 bits for the relocation type, but
currenlty we only define lower 8 bits due to the compatibility with
ELF32, so the idea is open those range to RV64, and add RV64 only
relocation to those range.

Also reserved part of relocation number to non-standard use.

We open up to 4096 and then still reserve the upper 20-bits for future
extension, which might use those bits other than identifying the
relocation type in the future, like SPARC64.

`RV64 only` might not precise enough, but enough for now since we don't have
RV64 with ilp32*, this should fix when adding that.