Relax gp could be platform specific register rather than reserved for linker relaxation
The usage of gp register has discussed serveral times (e.g. #298), and it
reserved as special register used for linker relaxation, that could be
improve perfomance and code size.

However it come with some limitation, like it can't applicable on shared
libraries, and it also not work well when program come with large datas
since the relaxable range is only +-2KiB.

Some platform like FreeBSD and Haiku never use gp in the whole system, and also
this might not useful in some baremetal system with specialized memory layout,
so we might consider to release the gp usage with a non-hard-ABI-breakage way.

Co-authored-by: Alex Bradbury <asb@asbradbury.org>