[Vector ABI] Type size and alignment for vector types.
The issue of Vector alignment is discussed in #347. It is mentioned that
aligning to 128 bytes might deliver better performance on some RISC-V cores,
but this behavior could lead to considerable stack wastage on zve32 and zve64
cores. For instance, in order to ensure a vector value in the stack conforms
to the ABI specification, we could potentially waste up to 96 bits per vector
object in stack for zve32, and the performance difference isn't always evident
across all core implementations.

Therefore, this proposal sets the alignment of vector types to element
alignment, to avoid wasting a significant amount of stack space in zve32 and
zve64 configurations. Also, since the ABI only specify the minimum alignment
and doesn't limit the compiler from adopting higher alignment for specific CPUs.

Fix #347.