Remove PT_RISCV_ATTRIBUTES
#71 added a PT_RISCV_ATTRIBUTES segment to contain the SHT_RISCV_ATTRIBUTES section. The
SHT_RISCV_ATTRIBUTES section does not have the SHF_ALLOC flag and is therefore not in a PT_LOAD
segment. Dynamic loaders cannot assume the PT_RISCV_ATTRIBUTES part is present (it may happen
to be mapped due to page alignment, but there is no guarantee).

To the best of my knowledge, PT_RISCV_ATTRIBUTES is the only open-source usage of a segment that
contains a non-SHF_ALLOC section.

I think the existence of PT_RISCV_ATTRIBUTES just complicates analysis tools and binary manipulation
tools like strip. E.g. if the SHT_RISCV_ATTRIBUTES section is removed, what should we do with
PT_RISCV_ATTRIBUTES?

I think we should remove PT_RISCV_ATTRIBUTES, but ensure that the program header type is not reused
by a future program header type.