Define contrainsts for ULEB128
Binutils [1] and a currently in-flight LLD patch [2] define ULEB128 such
that a SUB_ULEB128 must be preceded by a matching SET_ULEB128. This
means that if the first SET_ULEB128 that comes before a given
SUB_ULEB128 does not have the same relocation address, the linkers will
fail. A SET_ULEB128 also cannot appear alone.

Define this behavior in psABI such that libraries that read relocations
are able to make this same assumption. This change enforces that SET_ULEB128
must always immediately preceed the associated SUB_ULEB128, and vice
versa.

[1] https://sourceware.org/git/?p=binutils-gdb.git;a=blob;f=bfd/elfnn-riscv.c;h=09aa7be225ef3cba23fefd2ee977b3556a51368c;hb=HEAD#l2506
[2] https://reviews.llvm.org/D142880

Signed-off-by: Charlie Jenkins <charlie@rivosinc.com>