Revert "Add GP-relative relocations"
This reverts commit d49e48097bcad410e7e6d96ebf9d94d6d41dc9c0.

Both the ePIC and FDPIC ABIs are expected to define relocation types
which can be used for both PC-relative and GP-relative data, and are
expected to define relaxations to optimize access to both types of data.
Explicit, compiler-generated relocations using GP are only useful in
cases where an object is known at compile time to be local and present
in the GP-relative segment.  The relaxations will result in GP-relative
accesses, but the relocations involved will be linker internal and need
not have allocated standard numbers.