Rework Nonstandard Relocations
The primary change here is to require that vendor identifier symbols are
Defined, STB_LOCAL, STT_NOTYPE, and zero-sized.

This also slightly reorganises what was the "Vendor Identifiers" section
to bring together all the paragraphs about the nonstandard relocations,
so they are in one place. This rewording is not intended to change the
meaning of what was written before.