Relocation Sequencing
The rules around paired and marker relocations are reasonably complex,
so this adds a note about what could be expected for a single offset.

This change also clarifies a slight difference between "paired"
relocations (where you must have both, and they work together), and
"associated" relocations which effectively modify or clarify the
behaviour of the relocations at a specific offset. I classify
R_RISCV_RELAX and R_RISCV_VENDOR as "associated" relocations, as in the
case of the former, R_RISCV_RELAX is not required, and in the case of
the latter, the same R_RISCV_VENDOR "marker" is used with the 64
nonstandard relocations.

The only new rule here is that if you are using paired relocations, then
if one of the relocations is using a specific vendor identifier, then
the other must use the same vendor identifier.