introduction.adoc: Document how to treat encodings reserved for extensions
We want to make it clear that software should not start to assume that
an encoding reserved for a specific extension will always be for that
extension. For example, if that extension does not get ratified in a
reasonable amount of time, or if it needs to change its ABI such that an
encoding is no longer appropriate, such encodings will be reclaimed and
able to be reused for entirely different purposes.