TLSDESC Relocs Each Require a RELAX
These relaxations are designed so the sequence is not necessarily
adjacent. In this case, each of the relocations needs to be marked as
relaxable, as each instruction is separately modified/removed.

This also removes some spurious `(symbol)` annotations on
`R_RISCV_RELAX` - the relaxation relocation does not point to the same
symbol as the underlying relocation.