Vendor Symbols cannot always be LOCAL
The requirement for Vendor symbols to be `STB_LOCAL` was added to
parallel the requirement that mapping symbols be local, and also
resolved some issues with these symbols conflicting with user
definitions (local symbols can have duplicate names).

However, unlike mapping symbols, Vendor Symbols may appear in Dynamic
Symbol Tables, because we may have dynamic vendor relocations. Dynamic
symbol tables conventionally only hold GLOBAL and WEAK definitions.

This does bring up the problem of how to deal with a dynamic symbol
which has a name that matches a vendor symbol. This might become a QoI
decision for the linker whether to emit an error gracefully.